calculate_github_issue_for_commenting.py added (incomplete)
